Botao funcional, com titulo dinamico. Abstraido funcoes para outros arquivos usando extensions
